feat: do not copy symbolic links
Looking at GitHub Actions logs, it appears that the contents of
symbolic links are copied as the site is being built, which means that
a complete copy of the datasets is being created in each directory.
Since the lessons don't refer to these files directly (and can use `..`
if they need to), mccole was modified so that symlinks are only copied
on demand. This PR rebuilds the site without the symlinks in `./docs`.